backend/requirements.txt
# Traffic Detection Backend Requirements

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# YOLO
ultralytics>=8.0.0

# PyTorch (for SSD)
torch>=2.0.0
torchvision>=0.15.0

# Utilities
pydantic>=2.0.0
